refactor(tree): tree-factory-types.ts -> use-tree-types.ts
