+++ packages/devui-vue/devui/tree/src/core/use-check.ts
@@ -1,5 +1,5 @@
 import { Ref } from 'vue';
-import { IInnerTreeNode, IUseCore } from './tree-factory-types';
+import { IInnerTreeNode, IUseCore } from './use-tree-types';
 
 export default function useCheck(data: Ref<IInnerTreeNode[]>, core: IUseCore) {
   const { setNodeValue, getNode } = core;
