+++ packages/devui-vue/devui/tree/src/core/use-operate.ts
@@ -1,5 +1,5 @@
 import { Ref } from 'vue';
-import { IInnerTreeNode, ITreeNode, IUseCore } from './tree-factory-types';
+import { IInnerTreeNode, ITreeNode, IUseCore } from './use-tree-types';
 
 export default function useOperate(data: Ref<IInnerTreeNode[]>, core: IUseCore) {
   console.log('useOperate:', data, data.value);
